Move restore to src/common/ai and compile in main library to fix Windows build
