+++ src/common/ai/restore.h
